refactor(test): split FIDO2 PIN retry hardware test into a PinRetry folder
The single 816-line hardware characterization file mixed test flow with a large
amount of supporting infrastructure. Split it into a dedicated PinRetry/ folder:

- RecordingConnection.cs: the observational IYubiKeyConnection decorator (records
  raw status word / CTAP status / non-sensitive response bytes; enforces the
  per-phase wrong-PIN transmission budget) plus CommandObservation.
- PinRetryEvidence.cs: the evidence collector (now owns collector bookkeeping via
  intent-named Record* methods) and diagnostics rendering, plus the observation
  record types.
- Fido2PinRetryHardwareTests.cs: the two tests and their small helpers, now
  focused on the test flow.

Behavior-preserving: identical command sequencing, safety caps, and assertions.
Renamed the class to Fido2PinRetryHardwareTests and updated comments for accuracy.
Verified: integration builds, unit regression 8/8 pass, and the hardware tests
skip cleanly without the opt-in env vars.
